2021.07.15(목)   * queryDsl fetch join 테스트   * 연관객체 select 방식 공부

+++ src/main/java/kr/co/platform/api/posts/domain/entity/Posts.java
@@ -62,7 +62,7 @@ public class Posts{
     private List<PostsComment> comment;
     
     @Builder
-    public Posts(Long id, String author, String title, String content, int memberId) {
+    public Posts(Long id, String author, String title, String content, Long memberId) {
         this.id = id;
         this.author = author;
         this.title = title;
